requirements.txt
aiosqlite==0.20.0
colorama==0.4.6
py-cord==2.5.0
discord_webhook==1.3.1
httpx==0.27.0
python-dotenv==1.0.1
